refactor: Refactor MilvusCollectionService to delete and insert data instead of upserting
This commit refactors the MilvusCollectionService class to change the way data is updated. Instead of using the upsert method, the existing entry is deleted by its ID and then the new data is inserted. This approach ensures that the data is properly updated without any conflicts. The code also includes some additional improvements such as flushing the collection after deleting and inserting data.